requirements.txt
asgiref==3.8.1
Django==5.1.4
django-ckeditor==6.7.2
django-js-asset==3.0.1
ngrok==1.4.0
pillow==11.0.0
sqlparse==0.5.3
tzdata==2024.2
whitenoise==6.8.2
requests==2.31.0
gunicorn==23.0.0
python-dotenv==1.0.1
psycopg2==2.9.10